@@ -8,40 +8,40 @@ class IndexController extends ChangeNotifier {
   static const int PREVIOUS = -1;
   static const int MOVE = 0;
 
-  Completer _completer;
+  Completer? _completer;
 
-  int index;
-  bool animation;
-  int event;
+  int? index;
+  bool? animation;
+  int? event;
 
-  Future move(int index, {bool animation: true}) {
-    this.animation = animation ?? true;
+  Future? move(int index, {bool animation: true}) {
+    this.animation = animation;
     this.index = index;
     this.event = MOVE;
     _completer = new Completer();
     notifyListeners();
-    return _completer.future;
+    return _completer?.future;
   }
 
-  Future next({bool animation: true}) {
+  Future? next({bool animation: true}) {
     this.event = NEXT;
-    this.animation = animation ?? true;
+    this.animation = animation;
     _completer = new Completer();
     notifyListeners();
-    return _completer.future;
+    return _completer?.future;
   }
 
-  Future previous({bool animation: true}) {
+  Future? previous({bool animation: true}) {
     this.event = PREVIOUS;
-    this.animation = animation ?? true;
+    this.animation = animation;
     _completer = new Completer();
     notifyListeners();
-    return _completer.future;
+    return _completer?.future;
   }
 
   void complete() {
-    if (!_completer.isCompleted) {
-      _completer.complete();
+    if (_completer != null && !_completer!.isCompleted) {
+      _completer!.complete();
     }
   }
 }
