@@ -2,6 +2,7 @@ library transformer_page_view;
 
 import 'package:flutter/widgets.dart';
 import 'package:transformer_page_view/index_controller.dart';
+
 export 'package:transformer_page_view/index_controller.dart';
 export 'package:transformer_page_view/parallax.dart';
 
@@ -59,16 +60,16 @@ class TransformInfo {
   final Axis scrollDirection;
 
   TransformInfo(
-      {this.index,
-      this.position,
-      this.width,
-      this.height,
-      this.activeIndex,
-      this.fromIndex,
-      this.forward,
-      this.done,
-      this.viewportFraction,
-      this.scrollDirection});
+      {this.index = 0,
+      this.position = 0,
+      this.width = 0,
+      this.height = 0,
+      this.activeIndex = 0,
+      this.fromIndex = 0,
+      this.forward = false,
+      this.done = false,
+      this.viewportFraction = 0,
+      this.scrollDirection = Axis.horizontal});
 }
 
 abstract class PageTransformer {
@@ -86,7 +87,7 @@ typedef Widget PageTransformerBuilderCallback(Widget child, TransformInfo info);
 class PageTransformerBuilder extends PageTransformer {
   final PageTransformerBuilderCallback builder;
 
-  PageTransformerBuilder({bool reverse: false, @required this.builder})
+  PageTransformerBuilder({bool reverse: false, required this.builder})
       : assert(builder != null),
         super(reverse: reverse);
 
@@ -106,11 +107,11 @@ class TransformerPageController extends PageController {
     bool keepPage = true,
     double viewportFraction = 1.0,
     this.loop: false,
-    this.itemCount,
+    this.itemCount = 0,
     this.reverse: false,
   }) : super(
             initialPage: TransformerPageController._getRealIndexFromRenderIndex(
-                initialPage ?? 0, loop, itemCount, reverse),
+                initialPage, loop, itemCount, reverse),
             keepPage: keepPage,
             viewportFraction: viewportFraction);
 
@@ -128,13 +129,13 @@ class TransformerPageController extends PageController {
     if (itemCount == 0) return 0;
     int renderIndex;
     if (loop) {
-      renderIndex = index - kMiddleValue;
+      renderIndex = index.toInt() - kMiddleValue;
       renderIndex = renderIndex % itemCount;
       if (renderIndex < 0) {
         renderIndex += itemCount;
       }
     } else {
-      renderIndex = index;
+      renderIndex = index.toInt();
     }
     if (reverse) {
       renderIndex = itemCount - renderIndex - 1;
@@ -148,7 +149,7 @@ class TransformerPageController extends PageController {
     if (position.maxScrollExtent == null || position.minScrollExtent == null) {
       page = 0.0;
     } else {
-      page = super.page;
+      page = super.page ?? 0.0;
     }
 
     return page;
@@ -185,7 +186,7 @@ class TransformerPageController extends PageController {
 
   static int _getRealIndexFromRenderIndex(
       num index, bool loop, int itemCount, bool reverse) {
-    int result = reverse ? (itemCount - index - 1) : index;
+    int result = reverse ? (itemCount - index.toInt() - 1) : index.toInt();
     if (loop) {
       result += kMiddleValue;
     }
@@ -197,7 +198,7 @@ class TransformerPageView extends StatefulWidget {
   /// Create a `transformed` widget base on the widget that has been passed to  the [PageTransformer.transform].
   /// See [TransformInfo]
   ///
-  final PageTransformer transformer;
+  final PageTransformer? transformer;
 
   /// Same as [PageView.scrollDirection]
   ///
@@ -205,7 +206,7 @@ class TransformerPageView extends StatefulWidget {
   final Axis scrollDirection;
 
   /// Same as [PageView.physics]
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// Set to false to disable page snapping, useful for custom scroll behavior.
   /// Same as [PageView.pageSnapping]
@@ -213,12 +214,12 @@ class TransformerPageView extends StatefulWidget {
 
   /// Called whenever the page in the center of the viewport changes.
   /// Same as [PageView.onPageChanged]
-  final ValueChanged<int> onPageChanged;
+  final ValueChanged<int>? onPageChanged;
 
-  final IndexedWidgetBuilder itemBuilder;
+  final IndexedWidgetBuilder? itemBuilder;
 
   // See [IndexController.mode],[IndexController.next],[IndexController.previous]
-  final IndexController controller;
+  final IndexController? controller;
 
   /// Animation duration
   final Duration duration;
@@ -226,7 +227,7 @@ class TransformerPageView extends StatefulWidget {
   /// Animation curve
   final Curve curve;
 
-  final TransformerPageController pageController;
+  final TransformerPageController? pageController;
 
   /// Set true to open infinity loop mode.
   final bool loop;
@@ -253,9 +254,9 @@ class TransformerPageView extends StatefulWidget {
   /// [itemBuilder] will be called only with indices greater than or equal to
   /// zero and less than [itemCount].
   TransformerPageView({
-    Key key,
-    this.index,
-    Duration duration,
+    Key? key,
+    this.index = 0,
+    Duration? duration,
     this.curve: Curves.ease,
     this.viewportFraction: 1.0,
     this.loop: false,
@@ -267,28 +268,28 @@ class TransformerPageView extends StatefulWidget {
     this.transformer,
     this.itemBuilder,
     this.pageController,
-    @required this.itemCount,
-  })  : assert(itemCount != null),
+    required this.itemCount,
+  })   : assert(itemCount != null),
         assert(itemCount == 0 || itemBuilder != null || transformer != null),
         this.duration =
             duration ?? new Duration(milliseconds: kDefaultTransactionDuration),
         super(key: key);
 
   factory TransformerPageView.children(
-      {Key key,
-      int index,
-      Duration duration,
+      {Key? key,
+      int index = 0,
+      Duration? duration,
       Curve curve: Curves.ease,
       double viewportFraction: 1.0,
       bool loop: false,
       Axis scrollDirection = Axis.horizontal,
-      ScrollPhysics physics,
+      ScrollPhysics? physics,
       bool pageSnapping = true,
-      ValueChanged<int> onPageChanged,
-      IndexController controller,
-      PageTransformer transformer,
-      @required List<Widget> children,
-      TransformerPageController pageController}) {
+      ValueChanged<int>? onPageChanged,
+      IndexController? controller,
+      PageTransformer? transformer,
+      required List<Widget> children,
+      TransformerPageController? pageController}) {
     assert(children != null);
     return new TransformerPageView(
       itemCount: children.length,
@@ -316,7 +317,10 @@ class TransformerPageView extends StatefulWidget {
   }
 
   static int getRealIndexFromRenderIndex(
-      {bool reverse, int index, int itemCount, bool loop}) {
+      {bool reverse = false,
+      int index = 0,
+      int itemCount = 0,
+      bool loop = false}) {
     int initPage = reverse ? (itemCount - index - 1) : index;
     if (loop) {
       initPage += kMiddleValue;
@@ -325,11 +329,11 @@ class TransformerPageView extends StatefulWidget {
   }
 
   static PageController createPageController(
-      {bool reverse,
-      int index,
-      int itemCount,
-      bool loop,
-      double viewportFraction}) {
+      {bool reverse = false,
+      int index = 0,
+      int itemCount = 0,
+      bool loop = false,
+      double viewportFraction = 0.0}) {
     return new PageController(
         initialPage: getRealIndexFromRenderIndex(
             reverse: reverse, index: index, itemCount: itemCount, loop: loop),
@@ -338,71 +342,81 @@ class TransformerPageView extends StatefulWidget {
 }
 
 class _TransformerPageViewState extends State<TransformerPageView> {
-  Size _size;
-  int _activeIndex;
-  double _currentPixels;
+  Size? _size;
+  int? _activeIndex;
+  double? _currentPixels;
   bool _done = false;
 
   ///This value will not change until user end drag.
-  int _fromIndex;
+  int? _fromIndex;
 
-  PageTransformer _transformer;
+  PageTransformer? _transformer;
 
-  TransformerPageController _pageController;
+  TransformerPageController? _pageController;
 
   Widget _buildItemNormal(BuildContext context, int index) {
-    int renderIndex = _pageController.getRenderIndexFromRealIndex(index);
-    Widget child = widget.itemBuilder(context, renderIndex);
+    int? renderIndex = _pageController?.getRenderIndexFromRealIndex(index);
+    Widget child = widget.itemBuilder != null
+        ? widget.itemBuilder!(context, renderIndex ?? 0)
+        : Container();
     return child;
   }
 
   Widget _buildItem(BuildContext context, int index) {
-    return new AnimatedBuilder(
-        animation: _pageController,
-        builder: (BuildContext c, Widget w) {
-          int renderIndex = _pageController.getRenderIndexFromRealIndex(index);
-          Widget child;
-          if (widget.itemBuilder != null) {
-            child = widget.itemBuilder(context, renderIndex);
-          }
-          if (child == null) {
-            child = new Container();
-          }
-          if (_size == null) {
-            return child ?? new Container();
-          }
+    if (_pageController != null) {
+      return new AnimatedBuilder(
+          animation: _pageController!,
+          builder: (BuildContext c, Widget? w) {
+            int renderIndex =
+                _pageController!.getRenderIndexFromRealIndex(index);
+            Widget? child;
+            if (widget.itemBuilder != null) {
+              child = widget.itemBuilder!(context, renderIndex);
+            }
+            if (child == null) {
+              child = new Container();
+            }
+            if (_size == null) {
+              return child;
+            }
 
-          double position;
+            double? position;
 
-          double page = _pageController.realPage;
+            double? page = _pageController?.realPage;
 
-          if (_transformer.reverse) {
-            position = page - index;
-          } else {
-            position = index - page;
-          }
-          position *= widget.viewportFraction;
+            if (_transformer != null && _transformer!.reverse) {
+              position = (page ?? 0) - index;
+            } else {
+              position = (index - (page ?? 0)).toDouble();
+            }
+            position = position * widget.viewportFraction;
 
-          TransformInfo info = new TransformInfo(
-              index: renderIndex,
-              width: _size.width,
-              height: _size.height,
-              position: position.clamp(-1.0, 1.0),
-              activeIndex:
-                  _pageController.getRenderIndexFromRealIndex(_activeIndex),
-              fromIndex: _fromIndex,
-              forward: _pageController.position.pixels - _currentPixels >= 0,
-              done: _done,
-              scrollDirection: widget.scrollDirection,
-              viewportFraction: widget.viewportFraction);
-          return _transformer.transform(child, info);
-        });
+            TransformInfo info = new TransformInfo(
+                index: renderIndex,
+                width: _size?.width ?? 0,
+                height: _size?.height ?? 0,
+                position: position.clamp(-1.0, 1.0),
+                activeIndex: _pageController!
+                    .getRenderIndexFromRealIndex(_activeIndex ?? 0),
+                fromIndex: _fromIndex ?? 0,
+                forward:
+                    _pageController!.position.pixels - (_currentPixels ?? 0) >=
+                        0,
+                done: _done,
+                scrollDirection: widget.scrollDirection,
+                viewportFraction: widget.viewportFraction);
+
+            return _transformer?.transform(child, info) ?? Container();
+          });
+    }
+    return Container();
   }
 
-  double _calcCurrentPixels() {
-    _currentPixels = _pageController.getRenderIndexFromRealIndex(_activeIndex) *
-        _pageController.position.viewportDimension *
-        widget.viewportFraction;
+  double? _calcCurrentPixels() {
+    _currentPixels =
+        (_pageController?.getRenderIndexFromRealIndex(_activeIndex ?? 0) ?? 0) *
+            (_pageController?.position.viewportDimension ?? 0) *
+            widget.viewportFraction;
 
     //  print("activeIndex:$_activeIndex , pix:$_currentPixels");
 
@@ -415,13 +429,13 @@ class _TransformerPageViewState extends State<TransformerPageView> {
         _transformer == null ? _buildItemNormal : _buildItem;
     Widget child = new PageView.builder(
       itemBuilder: builder,
-      itemCount: _pageController.getRealItemCount(),
+      itemCount: _pageController?.getRealItemCount(),
       onPageChanged: _onIndexChanged,
       controller: _pageController,
       scrollDirection: widget.scrollDirection,
       physics: widget.physics,
       pageSnapping: widget.pageSnapping,
-      reverse: _pageController.reverse,
+      reverse: _pageController?.reverse ?? false,
     );
     if (_transformer == null) {
       return child;
@@ -445,18 +459,19 @@ class _TransformerPageViewState extends State<TransformerPageView> {
 
   void _onIndexChanged(int index) {
     _activeIndex = index;
-    if (widget.onPageChanged != null) {
-      widget.onPageChanged(_pageController.getRenderIndexFromRealIndex(index));
+    if (widget.onPageChanged != null && _pageController != null) {
+      widget
+          .onPageChanged!(_pageController!.getRenderIndexFromRealIndex(index));
     }
   }
 
   void _onGetSize(_) {
-    Size size;
+    Size? size;
     if (context == null) {
       onGetSize(size);
       return;
     }
-    RenderObject renderObject = context.findRenderObject();
+    RenderObject? renderObject = context.findRenderObject();
     if (renderObject != null) {
       Rect bounds = renderObject.paintBounds;
       if (bounds != null) {
@@ -467,13 +482,12 @@ class _TransformerPageViewState extends State<TransformerPageView> {
     onGetSize(size);
   }
 
-  void onGetSize(Size size) {
-    if(mounted){
+  void onGetSize(Size? size) {
+    if (mounted) {
       setState(() {
         _size = size;
       });
     }
-
   }
 
   @override
@@ -487,15 +501,15 @@ class _TransformerPageViewState extends State<TransformerPageView> {
           itemCount: widget.itemCount,
           loop: widget.loop,
           reverse:
-              widget.transformer == null ? false : widget.transformer.reverse);
+              widget.transformer == null ? false : widget.transformer!.reverse);
     }
     // int initPage = _getRealIndexFromRenderIndex(index);
     // _pageController = new PageController(initialPage: initPage,viewportFraction: widget.viewportFraction);
-    _fromIndex = _activeIndex = _pageController.initialPage;
+    _fromIndex = _activeIndex = _pageController?.initialPage;
 
     _controller = getNotifier();
     if (_controller != null) {
-      _controller.addListener(onChangeNotifier);
+      _controller!.addListener(onChangeNotifier);
     }
     super.initState();
   }
@@ -503,7 +517,7 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   @override
   void didUpdateWidget(TransformerPageView oldWidget) {
     _transformer = widget.transformer;
-    int index = widget.index ?? 0;
+    int index = widget.index;
     bool created = false;
     if (_pageController != widget.pageController) {
       if (widget.pageController != null) {
@@ -516,28 +530,29 @@ class _TransformerPageViewState extends State<TransformerPageView> {
             loop: widget.loop,
             reverse: widget.transformer == null
                 ? false
-                : widget.transformer.reverse);
+                : widget.transformer!.reverse);
       }
     }
 
-    if (_pageController.getRenderIndexFromRealIndex(_activeIndex) != index) {
-      _fromIndex = _activeIndex = _pageController.initialPage;
+    if (_pageController!.getRenderIndexFromRealIndex(_activeIndex ?? 0) !=
+        index) {
+      _fromIndex = _activeIndex = _pageController!.initialPage;
       if (!created) {
-        int initPage = _pageController.getRealIndexFromRenderIndex(index);
-        _pageController.animateToPage(initPage,
+        int initPage = _pageController!.getRealIndexFromRenderIndex(index);
+        _pageController!.animateToPage(initPage,
             duration: widget.duration, curve: widget.curve);
       }
     }
     if (_transformer != null)
-      WidgetsBinding.instance.addPostFrameCallback(_onGetSize);
+      WidgetsBinding.instance!.addPostFrameCallback(_onGetSize);
 
     if (_controller != getNotifier()) {
       if (_controller != null) {
-        _controller.removeListener(onChangeNotifier);
+        _controller!.removeListener(onChangeNotifier);
       }
       _controller = getNotifier();
       if (_controller != null) {
-        _controller.addListener(onChangeNotifier);
+        _controller!.addListener(onChangeNotifier);
       }
     }
     super.didUpdateWidget(oldWidget);
@@ -546,17 +561,17 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   @override
   void didChangeDependencies() {
     if (_transformer != null)
-      WidgetsBinding.instance.addPostFrameCallback(_onGetSize);
+      WidgetsBinding.instance!.addPostFrameCallback(_onGetSize);
     super.didChangeDependencies();
   }
 
-  ChangeNotifier getNotifier() {
+  ChangeNotifier? getNotifier() {
     return widget.controller;
   }
 
   int _calcNextIndex(bool next) {
-    int currentIndex = _activeIndex;
-    if (_pageController.reverse) {
+    int currentIndex = _activeIndex ?? 0;
+    if (_pageController != null && _pageController!.reverse) {
       if (next) {
         currentIndex--;
       } else {
@@ -570,11 +585,11 @@ class _TransformerPageViewState extends State<TransformerPageView> {
       }
     }
 
-    if (!_pageController.loop) {
-      if (currentIndex >= _pageController.itemCount) {
+    if (_pageController != null && !_pageController!.loop) {
+      if (currentIndex >= _pageController!.itemCount) {
         currentIndex = 0;
       } else if (currentIndex < 0) {
-        currentIndex = _pageController.itemCount - 1;
+        currentIndex = _pageController!.itemCount - 1;
       }
     }
 
@@ -582,42 +597,44 @@ class _TransformerPageViewState extends State<TransformerPageView> {
   }
 
   void onChangeNotifier() {
-    int event = widget.controller.event;
-    int index;
-    switch (event) {
-      case IndexController.MOVE:
-        {
-          index = _pageController
-              .getRealIndexFromRenderIndex(widget.controller.index);
-        }
-        break;
-      case IndexController.PREVIOUS:
-      case IndexController.NEXT:
-        {
-          index = _calcNextIndex(event == IndexController.NEXT);
-        }
-        break;
-      default:
-        //ignore this event
-        return;
-    }
-    if (widget.controller.animation) {
-      _pageController
-          .animateToPage(index,
-              duration: widget.duration, curve: widget.curve ?? Curves.ease)
-          .whenComplete(widget.controller.complete);
-    } else {
-      _pageController.jumpToPage(index);
-      widget.controller.complete();
+    if (widget.controller != null) {
+      int? event = widget.controller!.event;
+      int? index;
+      switch (event) {
+        case IndexController.MOVE:
+          {
+            index = _pageController
+                ?.getRealIndexFromRenderIndex(widget.controller!.index ?? 0);
+          }
+          break;
+        case IndexController.PREVIOUS:
+        case IndexController.NEXT:
+          {
+            index = _calcNextIndex(event == IndexController.NEXT);
+          }
+          break;
+        default:
+          //ignore this event
+          return;
+      }
+      if (widget.controller!.animation ?? false) {
+        _pageController
+            ?.animateToPage(index ?? 0,
+                duration: widget.duration, curve: widget.curve)
+            .whenComplete(widget.controller!.complete);
+      } else {
+        _pageController!.jumpToPage(index ?? 0);
+        widget.controller!.complete();
+      }
     }
   }
 
-  ChangeNotifier _controller;
+  ChangeNotifier? _controller;
 
   void dispose() {
     super.dispose();
     if (_controller != null) {
-      _controller.removeListener(onChangeNotifier);
+      _controller!.removeListener(onChangeNotifier);
     }
   }
 }
